requirements.txt
fastapi[standard]
sqlalchemy
pillow
